MAJOR: letsencryptforhaproxy: use of acme.sh '--ecc' option
first version of this script was written with an old version of acme.sh.
Newest version of acme.sh handles properly different directories for
both rsa and ecc certificates. IE: mycert and mycert_ecc for
respectively RSA and ECC.

So this patch cleans up the script to follow up those changes